requirements.txt
click==6.7
future==0.16.0
oauthlib==2.0.2
python-dotenv==0.6.4
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.8.0
